12 - Finalizes ProductsCheckout component
