+++ src/styles/ProductsCheckout.css
@@ -1,3 +1,12 @@
+.checkout-container {
+  border: 1px solid #ccc;
+  border-radius: 10px;
+  display: flex;
+  flex-direction: column;
+  margin: 20px;
+  padding: 20px;
+}
+
 .checkout-product {
   align-items: center;
   border: 1px solid #e5e5e5;
